refactor: 클래스, 메서드 네이밍 변경
