+++ README.md
@@ -5,5 +5,4 @@
 * 숫자 두개 이상을 컴마(,) 구분자로 입력할 경우 모든 숫자의 합을 반환하는 기능
 * 구분자로 콜론(:)도 사용할 수 있다
 * “//”와 “\n” 문자 사이에 커스텀 구분자를 지정할 수 있는 기능
-* 잘못된 구분자로 들어온 경우 IllegalArgumentException 이 발생
 * 음수를 전달할 경우 RuntimeException 이 발생
